requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
mysql-connector-python>=8.2.0
jinja2>=3.1.0
python-multipart>=0.0.6
